backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# Database
sqlalchemy>=2.0.38,<2.1

# Image processing
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Utils
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
httpx>=0.25.0,<0.28.0
cryptography>=41.0.0

# Logging
loguru==0.7.2
